# Use `ops.rsqrt`, improve normalization layers and enable ops fusion in tflite