--- a/keras_core/backend/numpy/math.py
+++ b/keras_core/backend/numpy/math.py
@@ -3,6 +3,7 @@ import numpy as np
 from keras_core.backend import standardize_dtype
 from keras_core.backend.jax.math import fft as jax_fft
 from keras_core.backend.jax.math import fft2 as jax_fft2
+from keras_core.backend.jax.math import rsqrt as jax_rsqrt
 from keras_core.backend.numpy.core import convert_to_tensor
 from keras_core.utils.module_utils import scipy
 
@@ -298,3 +299,7 @@ def istft(
     else:
         end = expected_output_len
     return x[..., start:end]
+
+
+def rsqrt(x):
+    return np.array(jax_rsqrt(x))