--- a/keras_core/layers/normalization/unit_normalization.py
+++ b/keras_core/layers/normalization/unit_normalization.py
@@ -45,7 +45,7 @@ class UnitNormalization(Layer):
         x = ops.cast(inputs, self.compute_dtype)
 
         square_sum = ops.sum(ops.square(x), axis=self.axis, keepdims=True)
-        x_inv_norm = 1 / ops.sqrt(ops.maximum(square_sum, 1e-12))
+        x_inv_norm = ops.rsqrt(ops.maximum(square_sum, 1e-12))
         return ops.multiply(x, x_inv_norm)
 
     def compute_output_shape(self, input_shape):