Fixes #824 

This PR accomplishes the following:
1. adding support for `rsqrt` in numpy backend (using jax's impl)
2. replacing `1 / ops.sqrt(x)` with `ops.rsqrt` for improved speed
3. reordering the ops in normalization layers to unify the implementation and match the expression of `tf.nn.batch_normalization` [link](https://github.com/tensorflow/tensorflow/blob/1cb1a030a62b169d90d34c747ab9b09f332bf905/tensorflow/python/ops/nn_impl.py#L1533)
4. Ensuring 100% unit test coverage for all normalization layers

After completing 3, tflite recognizes the pattern of CONV+BN+ReLU, and the ops are fused successfully.


<details>
<summary> standalone MobileNetV3 export script </summary>

```python
import tensorflow as tf

from keras_core.applications.mobilenet_v3 import MobileNetV3Small

keras_core_model = MobileNetV3Small(
    input_shape=(224, 224, 3), minimalistic=True
)

tf_callable = tf.function(
    keras_core_model.call,
    input_signature=[tf.TensorSpec((1, 224, 224, 3), tf.float32)],
    autograph=True,
    jit_compile=True,
)
tf_concrete_function = tf_callable.get_concrete_function()
converter = tf.lite.TFLiteConverter.from_concrete_functions(
    [tf_concrete_function], tf_callable
)
converter.optimizations = [tf.lite.Optimize.DEFAULT]
tflite_model = converter.convert()
with open("model.tflite", "wb") as f:
    f.write(tflite_model)

```

</details>

The visualization from netron: (before this PR vs. after this PR)
![](https://user-images.githubusercontent.com/20734616/262847739-dceba686-ce2d-48d6-9110-900fe5142fa8.jpg) ![圖片](https://github.com/keras-team/keras-core/assets/20734616/a43e556b-2a55-4dc6-8a1b-6acf38e9e724)

<details>
<summary> benchmark script </summary>

```python
from keras_core import layers
from keras_core import mixed_precision
from keras_core import models
from keras_core import ops

# "float32"
# "mixed_float16"
# "mixed_bfloat16"
dtype_policy = "float32"
mixed_precision.set_dtype_policy(dtype_policy)

x_train = ops.random.uniform(shape=(512, 64, 64, 64))
y_train = ops.random.uniform(shape=(512, 64, 64, 64))

# layers.BatchNormalization
# layers.GroupNormalization
# layers.LayerNormalization
normalization_cls = layers.LayerNormalization
normalization_args = {}
if normalization_cls is layers.GroupNormalization:
    normalization_args = {"groups": -1}

model = models.Sequential(
    [
        layers.InputLayer(shape=(64, 64, 64)),
        normalization_cls(**normalization_args),
        normalization_cls(**normalization_args),
        normalization_cls(**normalization_args),
    ]
)
model.compile(loss="mse", optimizer="adam")
model.fit(x_train, y_train, batch_size=128, epochs=3)

```

</details>

And the improvement:
|backend|layer|before this PR|after this PR|
|-|-|-|-|
|tensorflow|BatchNormalization|48ms/step|46ms/step|
|jax|BatchNormalization|49ms/step|48ms/step|
|torch|BatchNormalization|127ms/step|127ms/step|
|tensorflow|GroupNormalization|50ms/step|49ms/step|
|jax|GroupNormalization|51ms/step|50ms/step|
|torch|GroupNormalization|129ms/step|129ms/step|
|tensorflow|LayerNormalization|54ms/step|53ms/step|
|jax|LayerNormalization|55ms/step|54ms/step|
|torch|LayerNormalization|165ms/step|122ms/step|